fix: move all array to footer.tsx
